% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{add_feature_names}
\alias{add_feature_names}
\title{Adds column to a data frame for feature names}
\usage{
add_feature_names(df, colnumbers_to_use, remove_wo = TRUE)
}
\arguments{
\item{df}{data frame with columns of MSDial}

\item{colnumbers_to_use}{Column number of the columns to combine.
E.g. Alignment.ID = 1, Average.Rt.min = 2, Average.Mz = 3, Metabolite.name = 4.}

\item{remove_wo}{If True and Metabolite.name is used,
then "w/o MS2:" will be removed in the combined name.}
}
\value{
data frame
}
\description{
Combines columns of the data frame for new feature names. Each item of the columns is
separated with _ . The new column is add at the second place of the data frame.
}
