% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_overview.R
\name{calc_max_intensity}
\alias{calc_max_intensity}
\title{plot ordered maximum intensity of measured features}
\usage{
calc_max_intensity(df, sample_cols, fn_out = NULL)
}
\arguments{
\item{df}{data frame with feature list after}

\item{sample_cols}{list of column names of samples without blanks and quality controls}

\item{fn_out}{optional export plot as png}
}
\value{
list of data frame (df), plot and subplot
}
\description{
function calculates the maximum intensity over all samples for every feature and plot
them with logarithmic y-axis and a subplot of the last 50% with line for 2, 5, 10% of the
non-zero values. These lines in the subplot define values (without NA) in the samples,
which will be lost by the threshold on the y-axis.
}
