% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_overview.R
\name{calc_RSD}
\alias{calc_RSD}
\title{Bar plot with relative standard deviation of quality controls}
\usage{
calc_RSD(df, list_of_columns, fn_out = NULL)
}
\arguments{
\item{df}{data frame}

\item{list_of_columns}{List of column names of data frame to calculate RSD}

\item{fn_out}{optional export of the plot as png}
}
\value{
list of rsd values, the plot and a subplot without the last bar ( > 100%)
}
\description{
Calculates the relative standard deviation (RSD %) of quality controls and
creates a bar plot of the ranges 0 - 10 %, 10 - 20 % etc. The bars are
grouped be the number of missing values in the quality controls.
}
