% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_overview.R
\name{count_missing_values}
\alias{count_missing_values}
\title{Bar plot with the number of missing values in QC}
\usage{
count_missing_values(df, list_of_columns, fn_out = NULL)
}
\arguments{
\item{df}{data frame}

\item{list_of_columns}{List of column names of data frame for counting missing values}

\item{fn_out}{optional export of the plot as png}
}
\value{
list of array with count of missing values and bar plot
}
\description{
Function counts missing values in the quality control (QC) or other samples and
show the results in a bar plot, that can be exported as png.
}
