% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_dataframe}
\alias{filter_dataframe}
\title{Filter out rows in data frame}
\usage{
filter_dataframe(
  df,
  colnames = c("max.Intensity", "QC.missVal", "QC.RSD"),
  thresholds = c(10^5, 0, 35),
  modes = c(">=", "<=", "<=")
)
}
\arguments{
\item{df}{data frame}

\item{colnames}{array of column names for the filtering process}

\item{thresholds}{thresholds for every column define in colnames to filter out rows of the data frame}

\item{modes}{choose >= or <=}
}
\value{
data frame
}
\description{
The parameter colnames, thresholds and modes define the drop of rows by condition. E.g. the first
entry of the default values stands for the condition df$max.Intensity >= 10^5, so all rows with smaller
values than the threshold will be filtered out.
}
