% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_matched}
\alias{filter_matched}
\title{Filter out rows of a data frame}
\usage{
filter_matched(df, list_of_colnames)
}
\arguments{
\item{df}{data frame}

\item{list_of_colnames}{columns must be logical, thus contain False or True,
e.g. MS.MS.assigned, RT.matched, m.z.matched or MS.MS.matched}
}
\value{
data frame
}
\description{
Filter out rows of a data frame
}
