% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_overview.R
\name{get_all_sample_cols}
\alias{get_all_sample_cols}
\title{Get column names of all samples}
\usage{
get_all_sample_cols(df, first_sample_col, number_of_samples)
}
\arguments{
\item{df}{data frame}

\item{first_sample_col}{name of the first column of samples}

\item{number_of_samples}{number of all sample incl. blanks and quality controls}
}
\value{
list of column names of samples incl. blanks and quality controls
}
\description{
Get column names of all samples
}
