% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_overview.R
\name{get_cols}
\alias{get_cols}
\title{Get column names of specific samples}
\usage{
get_cols(all_sample_cols, target_word = NULL)
}
\arguments{
\item{all_sample_cols}{list of all sample columns in data frame}

\item{target_word}{Pattern to get specific columns, e.g. "Blank" or "QC" for quality control.
Use NULL, to return an empty list}
}
\value{
list of column names
}
\description{
Get column names of specific samples
}
