% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipids.R
\name{lipid_names_correction}
\alias{lipid_names_correction}
\title{Correction of lipid names in data frame}
\usage{
lipid_names_correction(lipid_df, with_summation = FALSE)
}
\arguments{
\item{lipid_df}{data frame with names in column 'Metabolite.name'}

\item{with_summation}{summation of number of carbon atoms and double bonds, e.g. x:y_a:b to x+a:y+b.
Default is False.}
}
\value{
data frame with new name column inserted
}
\description{
If the metabolites in a data frame are lipids, the names
will be corrected and insert as new column next to
metabolite.name. w/o MS2: and everything after | will be removed.
Summation of number of carbon atoms and double bonds is optional.
}
