% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_overview.R
\name{load_data}
\alias{load_data}
\title{Load feature list after preprocessing with MSDial}
\usage{
load_data(
  filepath,
  skip = 4,
  sep = "\\t",
  drop_last_cols = 2,
  drop_cols = c("MS1.isotopic.spectrum", "MS.MS.spectrum"),
  dec = "."
)
}
\arguments{
\item{filepath}{path of a csv or txt file}

\item{skip}{number of rows to skip before the data starts. Default value is 4 (for data of MSDial).}

\item{sep}{separator of the columns. Default value is a tabstopp (for data of MSDial).}

\item{drop_last_cols}{number of last columns in the data frame to drop. Default value is 2.}

\item{drop_cols}{names of columns to delete}

\item{dec}{decimal separator}
}
\value{
a data frame
}
\description{
Loads a csv or txt file as a data frame and drops the columns
MS1.isotopic.spectrum and MS.MS.spectrum.
}
