% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{normalize_intensity}
\alias{normalize_intensity}
\title{Normalize values of sample columns}
\usage{
normalize_intensity(df, df_norm_factors, decimals = 6)
}
\arguments{
\item{df}{data frame, that contains columns with samples}

\item{df_norm_factors}{data frame, that contains in first column part of sample names, to
address the columns in data frame df. The second column contains the factors for division.}

\item{decimals}{numbers of decimals for rounding}
}
\value{
data frame
}
\description{
Function divides values of sample columns, e.g. to normalize them by weight of samples.
For that, a data frame contains different target words for each sample and the factor for
division.
}
