% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{plot_max_intensity}
\alias{plot_max_intensity}
\title{Logarithmic plot of maximum intensities}
\usage{
plot_max_intensity(intensity_array)
}
\arguments{
\item{intensity_array}{Array of maximum intensities calculated by the function calc_max_intensity}
}
\value{
Plot
}
\description{
Logarithmic plot of maximum intensities
}
