% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{plot_missing_values}
\alias{plot_missing_values}
\title{Bar plot of missing values}
\usage{
plot_missing_values(missing_values_array)
}
\arguments{
\item{missing_values_array}{Array with missing values counted by the function count_missing_values}
}
\value{
bar plot
}
\description{
Bar plot of missing values
}
