% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{plot_rsd}
\alias{plot_rsd}
\title{Bar plot of relative standard deviation (rsd)}
\usage{
plot_rsd(rsd_array)
}
\arguments{
\item{rsd_array}{Array of rsd, that was calculated by the function calc_RSD}
}
\value{
bar plot
}
\description{
Bar plot of relative standard deviation (rsd)
}
