% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_overview.R
\name{remove_neg_mobility}
\alias{remove_neg_mobility}
\title{Remove -1 from data frame with ion mobility}
\usage{
remove_neg_mobility(df)
}
\arguments{
\item{df}{feature list exported from MS-Dial with CCS values}
}
\value{
data frame without rows of negative CCS values
}
\description{
Remove -1 from data frame with ion mobility
}
