% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_overview.R
\name{replace_missing_values}
\alias{replace_missing_values}
\title{replace negative intensities or zero with NA}
\usage{
replace_missing_values(df, all_sample_cols)
}
\arguments{
\item{df}{data frame (e.g. the data frame,
that function load_data of the NTAWorkflow Package returns)}

\item{all_sample_cols}{list of column names of all samples}
}
\value{
data frame
}
\description{
If there are negative intensities or zero in the samples, QC or blanks,
this function will replace them with NA (not available, so they are missing values).
}
