% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{transpose_df}
\alias{transpose_df}
\title{Exports data frame for further processing with MetaboAnalyst or SIMCA}
\usage{
transpose_df(
  df,
  colnum_of_name,
  all_sample_cols,
  sort_out = NULL,
  classes = NULL,
  remove_X = TRUE
)
}
\arguments{
\item{df}{data frame}

\item{colnum_of_name}{Number of the column in the data frame to use for feature names.}

\item{all_sample_cols}{list of column names of all samples}

\item{sort_out}{List of target words to sort out quality control (e.g. target word is "QC") or blanks}

\item{classes}{list of strings for grouping the samples}

\item{remove_X}{RStudio put an X before the name of a column, if it begins with a number.
After transpose of the matrix, X will be remove.}
}
\value{
data frame
}
\description{
Function creates new name for the features, takes the columns of samples, transposes the data frame and adds
a column with class names to group the samples. At least the data frame can be saved as a text file.
}
